requirements.txt
amqp==2.5.0
APScheduler==3.6.0
astroid==2.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2018.11.29
chardet==3.0.4
cloudinary==1.16.0
coverage==4.5.2
coveralls==1.5.1
dj-database-url==0.5.0
Django==2.2.2
django-cors-headers==2.4.0
django-heroku==0.3.1
django-nose==1.4.6
djangorestframework==3.9.4
djangorestframework-camel-case==0.2.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
gunicorn==19.9.0
idna==2.8
isort==4.3.4
kombu==4.6.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==3.0.5
nose==1.3.7
psycopg2==2.7.6.1
PyJWT==1.7.0
pylint==2.2.2
python-http-client==3.1.0
pytz==2019.1
redis==3.2.1
requests==2.21.0
sendgrid==6.0.5
six==1.12.0
sqlparse==0.3.0
typed-ast==1.1.1
tzlocal==1.5.1
urllib3==1.24.1
vine==1.3.0
whitenoise==4.1.2
wrapt==1.10.11
